simplified testbench

+++ hardware/simulation/cocotb/system_tb.py
@@ -95,20 +95,13 @@ async def files_tb_test(dut):
                 soc2cnsl.close()
                 break
             aux = cnsl2soc.read(1)
-            while(aux!=b''):
+            if(aux!=b''):
+                #print(aux, end = '')
+                #print('ENTER!')
                 send = int.from_bytes(aux, "little")
-                #print(chr(send), end = '')
-                number_of_bytes_from_cnsl += 1
-                if(number_of_bytes_from_cnsl%1000 == 0):
-                    print('-', end = '')
-                    sys.stdout.flush()
                 await uartwrite(dut, UART_TXDATA_ADDR, send)
-                TXready = 0
-                while(not TXready):
-                    TXready = await uartread(dut, UART_TXREADY_ADDR)
-                aux = cnsl2soc.read(1)
-            cnsl2soc.seek(0) # absolute file positioning
-            cnsl2soc.truncate() # to erase all data
+                cnsl2soc.seek(0) # absolute file positioning
+                cnsl2soc.truncate() # to erase all data
             cnsl2soc.close()
 
     print('TESTBENCH: finished\n\n')
